feat: added author verification

+++ communities/router.py
@@ -2,8 +2,10 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import select, insert
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from auth.models import User
 from communities.models import Community
 from communities.schemas import CreateCommunity, UpdateCommunity
+from dependencies import current_user
 from settings import get_async_session
 
 router = APIRouter(
@@ -45,8 +47,15 @@ async def get_community(community_id: int, session: AsyncSession = Depends(get_a
 
 
 @router.post("/create/", summary="Создать сообщество")
-async def create_community(new_community: CreateCommunity, session: AsyncSession = Depends(get_async_session)):
-    stmt = insert(Community).values(**new_community.dict())
+async def create_community(
+    new_community: CreateCommunity,
+    current_user: User = Depends(current_user),
+    session: AsyncSession = Depends(get_async_session)
+):
+    community_data = new_community.dict()
+    community_data["creator_id"] = current_user.id
+
+    stmt = insert(Community).values(**community_data)
     await session.execute(stmt)
     await session.commit()
     return {"status": "Created"}
@@ -56,6 +65,7 @@ async def create_community(new_community: CreateCommunity, session: AsyncSession
 async def update_community(
         community_id: int,
         community_data: UpdateCommunity,
+        current_user: User = Depends(current_user),
         session: AsyncSession = Depends(get_async_session),
 ):
     query = select(Community).where(Community.id == community_id)
@@ -64,6 +74,9 @@ async def update_community(
 
     if not existing_community:
         raise HTTPException(status_code=404, detail="Сообщество не найдено")
+
+    if existing_community.creator_id != current_user.id:
+        raise HTTPException(status_code=403, detail="У вас недостаточно прав для изменения этого сообщества")
 
     existing_community.name = community_data.name
     existing_community.description = community_data.description
@@ -78,6 +91,7 @@ async def update_community(
 @router.delete("/delete/{community_id}/", summary="Удалить сообщество")
 async def delete_community(
         community_id: int,
+        current_user: User = Depends(current_user),
         session: AsyncSession = Depends(get_async_session)
 ):
     query = select(Community).where(Community.id == community_id)
@@ -86,6 +100,9 @@ async def delete_community(
 
     if not existing_community:
         raise HTTPException(status_code=404, detail="Сообщество не найдено")
+
+    if existing_community.creator_id != current_user.id:
+        raise HTTPException(status_code=403, detail="У вас недостаточно прав для удаления этого сообщества")
 
     await session.delete(existing_community)
     await session.commit()
